requirements.txt
# requirements.txt
simpy>=4.0.2
numpy>=1.26.0
pyyaml>=6.0.1
matplotlib>=3.8.0
pandas>=2.1.0

# For plotting results if needed
matplotlib==3.7.1

# For data analysis
pandas==2.0.2 